+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Missions can now have new kinds of questions : define a color, number or free text for a particular tag
 
+### Changed
+- In review page, map is now wider
+
 
 ## 0.5.3 - 2018-10-28